Upadated tests to take into account change in types
git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/trunk/cdk@8495 eb4e18e3-b210-0410-a6ab-dec725e4b171

+++ src/org/openscience/cdk/test/AtomTypeTest.java
@@ -155,7 +155,7 @@ public class AtomTypeTest extends CDKTestCase {
 
         IAtomType a = builder.newAtomType("C");
         a.setValency(valency);
-        assertEquals(valency, a.getValency());
+        assertEquals(valency, (int) a.getValency());
     }
     public void testGetValency() {
         testSetValency_int();
@@ -166,7 +166,7 @@ public class AtomTypeTest extends CDKTestCase {
 
         IAtomType a = builder.newAtomType("C");
         a.setFormalNeighbourCount(count);
-        assertEquals(count, a.getFormalNeighbourCount());
+        assertEquals(count, (int) a.getFormalNeighbourCount());
     }
     public void testGetFormalNeighbourCount() {
         testSetFormalNeighbourCount_int();
@@ -330,7 +330,7 @@ public class AtomTypeTest extends CDKTestCase {
         IAtomType clone = (IAtomType)at.clone();
         
         at.setFormalNeighbourCount(2);
-        assertEquals(1, clone.getFormalNeighbourCount());
+        assertEquals(1, (int) clone.getFormalNeighbourCount());
     }
     
     /**
